+++ tests/Unit/inc/Engine/CriticalPath/Admin/Settings/addAsyncCssMobileOption.php
@@ -1,44 +1,36 @@
 <?php
 
-namespace WP_Rocket\Tests\Unit\inc\Engine\CriticalPath\AdminSubscriber;
+namespace WP_Rocket\Tests\Unit\inc\Engine\CriticalPath\Admin\Settings;
 
 use Brain\Monkey\Functions;
 use Mockery;
 use WP_Rocket\Admin\Options_Data;
 use WP_Rocket\Engine\Admin\Beacon\Beacon;
 use WP_Rocket\Engine\CriticalPath\CriticalCSS;
-use WP_Rocket\Engine\CriticalPath\AdminSubscriber;
+use WP_Rocket\Engine\CriticalPath\Admin\Settings;
 use WP_Rocket\Tests\Unit\TestCase;
 
 /**
- * @covers \WP_Rocket\Engine\CriticalPath\AdminSubscriber::add_async_css_mobile_option
+ * @covers \WP_Rocket\Engine\CriticalPath\Admin\Settings::add_async_css_mobile_option
  *
  * @group  CriticalPath
+ * @group  CriticalPathSettings
  */
 class Test_AddAsyncCssMobileOption extends TestCase {
-
-	public function setUp() {
-		parent::setUp();
-
-		Functions\when( 'get_current_blog_id' )->justReturn( 1 );
-	}
-
 	/**
 	 * @dataProvider configTestData
 	 */
 	public function testShouldAddOption( $options, $expected ) {
-		$subscriber = new AdminSubscriber(
+		$settings = new Settings(
 			Mockery::mock( Options_Data::class ),
 			Mockery::mock( Beacon::class ),
 			Mockery::mock( CriticalCSS::class ),
-			'wp-content/cache/critical-css/',
 			'wp-content/plugins/wp-rocket/views/metabox/cpcss'
 		);
 
-
 		$this->assertSame(
 			$expected,
-			$subscriber->add_async_css_mobile_option( $options )
+			$settings->add_async_css_mobile_option( $options )
 		);
 	}
 }
