+++ tests/Unit/inc/Engine/CriticalPath/Admin/Settings/enableMobileCpcss.php
@@ -1,26 +1,36 @@
 <?php
 
-namespace WP_Rocket\Tests\Unit\inc\Engine\CriticalPath\AdminSubscriber;
+namespace WP_Rocket\Tests\Unit\inc\Engine\CriticalPath\Admin\Settings;
 
 use Brain\Monkey\Functions;
+use Mockery;
+use WP_Rocket\Admin\Options_Data;
+use WP_Rocket\Engine\Admin\Beacon\Beacon;
+use WP_Rocket\Engine\CriticalPath\Admin\Settings;
+use WP_Rocket\Engine\CriticalPath\CriticalCSS;
 use WP_Rocket\Tests\Unit\TestCase;
 
 /**
- * @covers \WP_Rocket\Engine\CriticalPath\AdminSubscriber::enable_mobile_cpcss
+ * @covers \WP_Rocket\Engine\CriticalPath\Admin\Settings::enable_mobile_cpcss
  * @uses   ::rocket_get_constant
  *
  * @group  CriticalPath
+ * @group  CriticalPathSettings
  */
 class Test_EnableMobileCpcss extends TestCase {
-	use GenerateTrait;
-
 	public function setUp() {
 		parent::setUp();
 
-		Functions\when( 'get_current_blog_id' )->justReturn( 1 );
 		Functions\when( 'check_ajax_referer' )->justReturn( true );
 
-		$this->setUpMocks();
+		$this->options      = Mockery::mock( Options_Data::class );
+		$this->critical_css = Mockery::mock( CriticalCSS::class );
+		$this->settings     = new Settings(
+			$this->options,
+			Mockery::mock( Beacon::class ),
+			$this->critical_css,
+			'wp-content/plugins/wp-rocket/views/cpcss'
+		);
 	}
 
 	/**
@@ -45,6 +55,6 @@ class Test_EnableMobileCpcss extends TestCase {
 			Functions\expect( 'wp_send_json_success' )->once();
 		}
 
-		$this->subscriber->enable_mobile_cpcss();
+		$this->settings->enable_mobile_cpcss();
 	}
 }
