+++ tests/Unit/inc/Engine/CriticalPath/Admin/Settings/setAsyncCssMobileDefaultValue.php
@@ -1,34 +1,28 @@
 <?php
 
-namespace WP_Rocket\Tests\Unit\inc\Engine\CriticalPath\AdminSubscriber;
+namespace WP_Rocket\Tests\Unit\inc\Engine\CriticalPath\Admin\Settings;
 
 use Brain\Monkey\Functions;
 use Mockery;
 use WP_Rocket\Admin\Options_Data;
 use WP_Rocket\Engine\Admin\Beacon\Beacon;
 use WP_Rocket\Engine\CriticalPath\CriticalCSS;
-use WP_Rocket\Engine\CriticalPath\AdminSubscriber;
+use WP_Rocket\Engine\CriticalPath\Admin\Settings;
 use WP_Rocket\Tests\Unit\TestCase;
 
 /**
- * @covers \WP_Rocket\Engine\CriticalPath\AdminSubscriber::set_async_css_mobile_default_value
+ * @covers \WP_Rocket\Engine\CriticalPath\Admin\Settings::set_async_css_mobile_default_value
  *
  * @group  CriticalPath
+ * @group  CriticalPathSettings
  */
 class Test_SetAsyncCssMobileDefaultValue extends TestCase {
-
-	public function setUp() {
-		parent::setUp();
-
-		Functions\when( 'get_current_blog_id' )->justReturn( 1 );
-	}
-
 	/**
 	 * @dataProvider configTestData
 	 */
 	public function testShouldUpdateOption( $versions, $update ) {
-		$options = Mockery::mock( Options_Data::class );
-		$subscriber = new AdminSubscriber(
+		$options  = Mockery::mock( Options_Data::class );
+		$settings = new Settings(
 			$options,
 			Mockery::mock( Beacon::class ),
 			Mockery::mock( CriticalCSS::class ),
@@ -37,7 +31,7 @@ class Test_SetAsyncCssMobileDefaultValue extends TestCase {
 		);
 
         if ( true === $update ) {
-            $settings = [
+            $options_value = [
                 'async_css_mobile' => 1,
             ];
 
@@ -47,15 +41,15 @@ class Test_SetAsyncCssMobileDefaultValue extends TestCase {
 
             $options->shouldReceive( 'get_options' )
                 ->once()
-                ->andReturn( $settings );
+                ->andReturn( $options_value );
 
             Functions\expect( 'update_option' )
                 ->once()
-                ->with( 'wp_rocket_settings', $settings );
+                ->with( 'wp_rocket_settings', $options_value );
         } else {
             Functions\expect( 'update_option' )->never();
         }
 
-        $subscriber->set_async_css_mobile_default_value( $versions['new'], $versions['old'] );
+        $settings->set_async_css_mobile_default_value( $versions['new'], $versions['old'] );
 	}
 }
